baseimportExport ccFunction js
