+++ blocks/form/functions.js
@@ -2,39 +2,39 @@
 import {
   getOTP,
   otpValidation,
-  customSetFocus,
-  journeyResponseHandler,
-  createJourneyId,
-  sendAnalytics,
-  resendOTP,
+  // customSetFocus,
+  // journeyResponseHandler,
+  // createJourneyId,
+  // sendAnalytics,
+  // resendOTP,
   hideLoaderGif,
-  validateLogin,
-  getAddressDetails,
-  pinCodeMaster,
-  validateEmailID,
-  currentAddressToggleHandler,
-  otpValHandler,
-  setNameOnCard,
-  prefillForm,
-  getThisCard,
+  // validateLogin,
+  // getAddressDetails,
+  // pinCodeMaster,
+  // validateEmailID,
+  // currentAddressToggleHandler,
+  // otpValHandler,
+  // setNameOnCard,
+  // prefillForm,
+  // getThisCard,
   validatePan,
   panAPISuccesHandler,
-  executeInterfaceApi,
-  executeInterfaceApiFinal,
-  ipaRequestApi,
-  ipaSuccessHandler,
-  executeInterfaceResponseHandler,
-  aadharConsent123,
-  documentUpload,
+  // executeInterfaceApi,
+  // executeInterfaceApiFinal,
+  // ipaRequestApi,
+  // ipaSuccessHandler,
+  // executeInterfaceResponseHandler,
+  // aadharConsent123,
+  // documentUpload,
   fetchAuthCode,
-  checkMode,
+  // checkMode,
   aadharInit,
   redirect,
   reloadPage,
   idcomUrlSet,
   idcomRedirection,
-  executeInterfacePostRedirect,
-  invokeJourneyDropOffUpdate, invokeJourneyDropOff, invokeJourneyDropOffByParam,
+  // executeInterfacePostRedirect,
+  // invokeJourneyDropOffUpdate, invokeJourneyDropOff, invokeJourneyDropOffByParam,
 } from '../../common/functions.js';
 
 import {
@@ -47,6 +47,34 @@ import {
   sendGenericClickEvent,
 } from '../../common/analytics.js';
 
+import {
+  executeInterfaceApiFinal,
+  executeInterfaceApi,
+  ipaRequestApi,
+  ipaSuccessHandler,
+  executeInterfacePostRedirect,
+  executeInterfaceResponseHandler,
+  journeyResponseHandler,
+  createJourneyId,
+  sendAnalytics,
+  resendOTP,
+  customSetFocus,
+  validateLogin,
+  getAddressDetails,
+  pinCodeMaster,
+  validateEmailID,
+  currentAddressToggleHandler,
+  otpValHandler,
+  setNameOnCard,
+  prefillForm,
+  getThisCard,
+  aadharConsent123,
+  invokeJourneyDropOff,
+  invokeJourneyDropOffByParam,
+  invokeJourneyDropOffUpdate,
+  documentUpload,
+  checkMode,
+} from '../../creditcards/corporate-creditcard/cc-functions.js';
 import { CURRENT_FORM_CONTEXT as currentFormContext } from '../../common/constants.js';
 
 /**
